+++ app_estudo/__init__.py
@@ -1,0 +1,1 @@
+"""Pacote base do projeto app_estudo."""