+++ app_estudo/domain/__init__.py
@@ -1,0 +1,5 @@
+"""Camada de dominio do projeto."""
+
+from .study_item import CuratedStudyItem
+
+__all__ = ["CuratedStudyItem"]